+++ openehr/src/base/mod.rs
@@ -18,12 +18,14 @@
 pub mod interval;
 pub mod iso8601;
 pub mod object_id;
+pub mod real;
 pub mod object_ref;
 #[doc(hidden)]
 pub mod serde_support;
 pub mod uid;
 
 pub use interval::{Interval, SemanticOrd};
+pub use real::Real;
 pub use iso8601::{Date, DatePrecision, DateTime, Duration, Offset, Time, TimePrecision};
 pub use object_id::{
     ArchetypeId, GenericId, HierObjectId, ObjectId, ObjectVersionId, TemplateId, TerminologyId,